+++ packages/api/tests/lib/reconciliationReport/test-normalizeEvent.js
@@ -1,20 +1,7 @@
 const test = require('ava');
+const omit = require('lodash/omit');
 const { InvalidArgument } = require('@cumulus/errors');
 const { normalizeEvent } = require('../../../lib/reconciliationReport/normalizeEvent');
-
-test('normalizeEvent throws error if array of collectionIds passed to Internal report', (t) => {
-  const inputEvent = {
-    systemBucket: 'systemBucket',
-    stackName: 'stackName',
-    startTimestamp: new Date().toISOString(),
-    endTimestamp: new Date().toISOString(),
-    reportType: 'Internal',
-    collectionId: ['someCollection___version'],
-  };
-  t.throws(() => normalizeEvent(inputEvent), {
-    message: 'collectionId: ["someCollection___version"] is not valid input for an \'Internal\' report.',
-  });
-});
 
 test('normalizeEvent converts input key collectionId string to length 1 array on collectionIds', (t) => {
   const inputEvent = {
@@ -22,7 +9,7 @@ test('normalizeEvent converts input key collectionId string to length 1 array on
     stackName: 'stackName',
     startTimestamp: new Date().toISOString(),
     endTimestamp: new Date().toISOString(),
-    reportType: 'NotInternal',
+    reportType: 'anytype',
     collectionId: 'someCollection___version',
   };
   const expect = { ...inputEvent, collectionIds: ['someCollection___version'] };
@@ -38,7 +25,7 @@ test('normalizeEvent moves input key collectionId array to array on collectionId
     stackName: 'stackName',
     startTimestamp: new Date().toISOString(),
     endTimestamp: new Date().toISOString(),
-    reportType: 'NotInternal',
+    reportType: 'anytype',
     collectionId: ['someCollection___version', 'secondcollection___version'],
   };
   const expect = {
@@ -51,31 +38,13 @@ test('normalizeEvent moves input key collectionId array to array on collectionId
   t.deepEqual(actual, expect);
 });
 
-test('normalizeEvent adds new collectionIds key when collectionId passed to Internal report', (t) => {
-  const inputEvent = {
-    systemBucket: 'systemBucket',
-    stackName: 'stackName',
-    startTimestamp: new Date().toISOString(),
-    endTimestamp: new Date().toISOString(),
-    reportType: 'Internal',
-    collectionId: 'someCollection___version',
-  };
-  const expect = {
-    ...inputEvent,
-    collectionIds: ['someCollection___version'],
-  };
-
-  const actual = normalizeEvent(inputEvent);
-  t.deepEqual(actual, expect);
-});
-
 test('normalizeEvent throws error if original input event contains collectionIds key', (t) => {
   const inputEvent = {
     systemBucket: 'systemBucket',
     stackName: 'stackName',
     startTimestamp: new Date().toISOString(),
     endTimestamp: new Date().toISOString(),
-    reportType: 'Internal',
+    reportType: 'anytype',
     collectionIds: ['someCollection___version'],
   };
   t.throws(() => normalizeEvent(inputEvent), {
@@ -89,7 +58,7 @@ test('normalizeEvent moves string on granuleId to array on granuleIds', (t) => {
     stackName: 'stackName',
     startTimestamp: new Date().toISOString(),
     endTimestamp: new Date().toISOString(),
-    reportType: 'Not Internal',
+    reportType: 'anytype',
     granuleId: 'someGranule',
   };
   const expect = {
@@ -108,7 +77,7 @@ test('normalizeEvent moves array on granuleId to granuleIds', (t) => {
     stackName: 'stackName',
     startTimestamp: new Date().toISOString(),
     endTimestamp: new Date().toISOString(),
-    reportType: 'Not Internal',
+    reportType: 'anytype',
     granuleId: ['someGranule', 'someGranule2'],
   };
 
@@ -121,21 +90,6 @@ test('normalizeEvent moves array on granuleId to granuleIds', (t) => {
   const actual = normalizeEvent(inputEvent);
   t.deepEqual(actual, expect);
 });
-
-test('normalizeEvent throws error if array of granuleIds is passed to Internal report', (t) => {
-  const inputEvent = {
-    systemBucket: 'systemBucket',
-    stackName: 'stackName',
-    startTimestamp: new Date().toISOString(),
-    endTimestamp: new Date().toISOString(),
-    reportType: 'Internal',
-    granuleId: ['someGranuleId'],
-  };
-  t.throws(() => normalizeEvent(inputEvent), {
-    message: 'granuleId: ["someGranuleId"] is not valid input for an \'Internal\' report.',
-  });
-});
-
 test('normalizeEvent throws error if granuleIds and collectionIds are passed to non-Internal report', (t) => {
   const inputEvent = {
     systemBucket: 'systemBucket',
@@ -165,7 +119,7 @@ test('normalizeEvent correctly handles granuleIds, collectionIds, and providers 
   };
 
   const expected = {
-    ...inputEvent,
+    ...omit(inputEvent, ['collectionId', 'granuleId']),
     granuleIds: ['someGranuleId'],
     collectionIds: ['someCollectionId1'],
     providers: ['someProvider1'],